Added GUI MCP server module and multi-day-data skill.
  -- Added gui_mcp_server.py with viewer lifecycle and data query tools for agentic GUI interaction.
  -- Added 'mcp' subcommand to cindra-gui CLI for starting the GUI MCP server.
  -- Replaced multi-day-config skill with multi-day-data skill combining configuration and data references.
  -- Updated multi-day-processing and single-day-data skills to reference the new multi-day-data skill.
  -- Fixed plane_index dtype from uint8 to int32 in ROI statistics serialization and skill documentation.

+++ src/cindra/dataclasses/single_day_data.py
@@ -874,7 +874,7 @@ class ROIStatistics:
             dtype=np.float32,
         )
 
-        plane_index = np.array([roi.plane_index for roi in roi_list], dtype=np.uint8)
+        plane_index = np.array([roi.plane_index for roi in roi_list], dtype=np.int32)
 
         save_dict: dict[str, np.ndarray] = {
             "footprints": footprints,
